Add support for providing alt text
To make published GIFs, via Micropub, more accessible, we should allow
alt text to be added to the images.

As per conversation on #3, this can't be reliably done through the
Gfycat API, so we should instead allow the user to provide this
information.

As we're using a JSON post, we need to construct an object for the
photo property as per https://micropub.spec.indieweb.org/#json-syntax

Closes #3.