Correctly use form-encoded Micropub requests
It looks like we've been using JSON encoded requests, which is the
default method for the micropub-helper library.

Form-encoded is generally easier for servers to implement, and as we've
advertised on the /preview/ page, so we should try and use that instead
of JSON.

This requires a slightly different interface to use, so we need to
restructure how we send the request.